move shutdown logic to nb/kernel/
